+++ src/enums.jl
@@ -2,6 +2,7 @@ import Base: *, sign
 using EnumX
 
 @enumx TradeDir::Int8 Null = 0 Buy = 1 Sell = -1
+@enumx SettlementStyle::Int8 Asset = 1 Cash = 2
 
 @inline sign(x::TradeDir.T) = Quantity(Int8(x))
 @inline is_long(dir::TradeDir.T) = dir == TradeDir.Buy