+++ src/instrument.jl
@@ -13,8 +13,10 @@ mutable struct Instrument{IData}
     const quote_symbol::Symbol
     const quote_tick::Price       # minimum price increment of base asset
     const quote_digits::Int       # number of digits after the decimal point for display
+
+    const settlement::SettlementStyle.T
     quote_cash_index::Int
- 
+
     const multiplier::Float64
     const metadata::IData
 
@@ -29,6 +31,7 @@ mutable struct Instrument{IData}
         base_digits=2,
         quote_tick::Price=0.01,
         quote_digits=2,
+        settlement::SettlementStyle.T=SettlementStyle.Cash,
         multiplier::Float64=1.0,
         metadata::IData=nothing
     ) where {IData}
@@ -43,6 +46,7 @@ mutable struct Instrument{IData}
             quote_symbol,
             quote_tick,
             quote_digits,
+            settlement,
             0, # quote_cash_index
             multiplier,
             metadata