+++ src/enums.jl
@@ -3,6 +3,7 @@ using EnumX
 
 @enumx TradeDir::Int8 Null = 0 Buy = 1 Sell = -1
 @enumx SettlementStyle::Int8 Asset = 1 Cash = 2
+@enumx MarginMode::Int8 None = 0 PercentNotional = 1 FixedPerContract = 2
 
 @inline sign(x::TradeDir.T) = Quantity(Int8(x))
 @inline is_long(dir::TradeDir.T) = dir == TradeDir.Buy